refactored python benchmark in separate modules
